feat: Add `Default` trait implementations and enhance code readability
- Introduce `impl Default` for multiple structs across architecture and IR modules to enable streamlined initialization.
- Shift from `or_insert_with` to `or_default`, simplifying BTreeMap usage across various analysis systems.
- Replace pattern-based matches with inline condition checks for better clarity (e.g., `if let` optimizations).
- Improve iterator handling by utilizing `.values()` and `.keys()` to enhance loop readability and efficiency.
- Perform minor cleanups in `fireball` modules, including unnecessary reference removal and method chaining standardization.

+++ fireball/src/binary/elf/mod.rs
@@ -87,7 +87,7 @@ impl Elf {
                     real_name: None,
                     virtual_address: phdr.p_vaddr,
                     virtual_size: phdr.p_memsz,
-                    file_offset: phdr.p_offset as u64,
+                    file_offset: phdr.p_offset,
                     size_of_file: phdr.p_filesz,
                 };
